feat(core) implement a dj projects shared DB POC
**DO NOT MERGE**

This work is a proof-of-concept of making Fonzie a standalone
application consumming Open edX data. Here is how we proceeded:

* add a fake Open edX django project running in docker containers (see
  docker-compose.yml); this is the project that bundles a students app
  with the Student model that will be exposed by Fonzie
* configure fonzie as a project using the same PostgreSQL database as
  the Open edX fake project (see test_settings.py)
* implement a serializer and an API view (in fonzie app) for the Student
  model

And... tada! We were able to expose Open edX data with Fonzie, without
having to integrate fonzie in an Open edX instance and without adding
Open edX django applications in our `INSTALLED_APPS`.

*Known constrain*: models MUST define an explicit `app_label`

That said, this work stays a POC, and does not take into account Open
edX platform black magic and complexity.

The next step is to use the fun-platform dockerized stack and expose
Open edX student.UserProfile model as a real world example.

Which me good luck!