🐛(tests) use LMS settings in tests
** FIX DETAILS **

* configure & test DRF API versioning
* use lms settings in tests
* get rid of tox in our CI:

Using tox within docker containers seems overkill; let's make things a
little bit "simpler" by dropping it once for all.

* Prevent root user creation:

when trying to build the base container with no user or group as build
arguments, docker tries to create the root user. We prevent this
behavior by only creating a user when uid:gid are distinct from root
user ids.

* Update volume endpoints:

fundocker/edxapp now use the same application tree as edx, we need to
stick to it to make it works.

** MISC **

* fix broken comparison in ci script
* use super() in test cases setUp method
* remove unused CMS service
* improve CI step switch/case