🐛(requirements) fix default requirements
Default requirements.txt file should not install development
dependencies. Development dependencies requirements file has been moved
to requirements-dev.txt. And we kept a minimalist requirements.txt to
maintain compatibility with tools requiring it at the root of the
project.

The Dockerfile now uses the development requirements as this image is
intended to get used only in development (for now).