👷(circleci) move from Travis to Circle
Now that we use to build our CI workflow with CircleCI and that we are
fine with it, we've decided to give Fonzie's CI a little refactoring.

The node stack has been removed since we only used it to install Dredd
that is now installed via a Docker compose service.