🐛(user) allow to get user info even if it is inactive
Currently, when user account is inactive as the user view is using the
default SessionAuthentication classes the request returns a 403
Forbidden. In other part of OpenEdX, user requests related to its own
account are allowed even if the account is inactive. That's why we
allowed a user to retrieve its own account information even if its
account is inactive.

Fix #32