✨(user) set cross domain csrf token on user get
OpenEdX has an API route to update username profile but this one
requires to send a cross domain CSRF Token. To ensure that the api
consumer has a valid CSRF Token, we decorate the method `get` of the
user api to add a `Set-Cookie` header to set the CSRF Token cookie on
request.